requirements.txt
PyMuPDF
numpy
pillow
openai
nltk
python-Levenshtein
python-dotenv
psycopg2
uvicorn
fastapi